Addressed Comments.

+++ src/main/java/com/studenthub/course/entity/Student.java
@@ -2,8 +2,10 @@ package com.studenthub.course.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.studenthub.course.validator.NoNumbers;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
@@ -14,7 +16,7 @@ import java.time.LocalDate;
 public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @NoNumbers
     private String name;
